+++ model/tag.go
@@ -14,9 +14,9 @@ type Tags []*Tag
 
 // Tag represents a tag model definition.
 type Tag struct {
-	ID        int       `json:"id" gorm:"primary_key"`
+	ID        int64     `json:"id" gorm:"primary_key"`
 	Repo      *Repo     `json:"repo,omitempty"`
-	RepoID    int       `json:"repo_id" sql:"index"`
+	RepoID    int64     `json:"repo_id" sql:"index"`
 	Slug      string    `json:"slug"`
 	Name      string    `json:"name"`
 	FullName  string    `json:"full_name"`
